requirements.txt
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
mem0ai>=1.0.0
lancedb>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.25.0
websockets>=12.0
openai>=1.0.0
twilio>=8.0.0
typing-extensions>=4.0.0

